First Creation
As part of my education in Python and GitHub, I’ve created this
Rock-Paper-Scissors game.

There would be a more efficient way to code the outcomes than what is
presented here. I’ve deliberately published this less efficient version
so that I may get some experience in the process of using branches and
updating scripts with GitHub.